move board dimension to origin, update ws2812 package, reroute
The lower left corner of the board dimension was not coincident with the
coordinates 0,0 which lead to difficulties moving an exported dxf to the
right place for drilling/milling, so I moved everything an half inch to
the right. Then I used the new centered ws2812 package and routed the
wires again. The drills are still in place, but next time the board is
slightly different.